% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashalpha.R
\docType{package}
\name{cashalpha-package}
\alias{cashalpha}
\alias{cashalpha-package}
\title{: Correlated Adaptive Shrinkage, using Empirical Bayes}
\description{
: The R package 'cashalpha' fits Empirical Bayes Normal Means with correlated noise using the exchangeable correlated noise model
}
\author{
: Lei Sun, Matthew Stephens, \email{sunl@uchicago.edu}
}
