% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdash_lik.R
\name{gdash}
\alias{gdash}
\title{Fit Empirical Bayes Normal Means with Correlated Noise}
\usage{
gdash(betahat, sebetahat, mixcompdist = "normal", method = "fdr",
  gd.normalized = TRUE, primal = FALSE, gd.ord = 10, w.lambda = 10,
  w.rho = 0.5, w.pen = NULL, gd.priority = FALSE, lfsr = FALSE,
  control = list(maxiter = 50))
}
\arguments{
\item{betahat}{A p vector of observations}

\item{sebetahat}{A p vector of known standard error.}
}
\description{
This is the main interface for fitting ECSN models
  based on algorithms proposed by Sun and Stephens.  The default
  behaviour is simply to run the biconvex optimization and return
  the result.
}
